iis setup is working now

+++ iis_setup.ps1
@@ -1,14 +1,26 @@
 ﻿#Install-WindowsFeature -name Web-Server -IncludeManagementTools
-Import-Module ServerManager
-Add-WindowsFeature Web-Scripting-Tools
 Import-Module WebAdministration
-$SiteFolderPath = "C:\WebSite"              # Website Folder
-$SiteAppPool = "MyAppPool"                  # Application Pool Name
-$SiteName = "MySite"                        # IIS Site Name
-$SiteHostName = "www.TestSite.com"          # Host Header
 
-New-Item $SiteFolderPath -type Directory
-Set-Content $SiteFolderPath\Default.htm "<h1>Hello IIS</h1>"
-New-Item IIS:\AppPools\$SiteAppPool
-New-Item IIS:\Sites\$SiteName -physicalPath $SiteFolderPath -bindings @{protocol="http";bindingInformation=":80:"+$SiteHostName}
-Set-ItemProperty IIS:\Sites\$SiteName -name applicationPool -value $SiteAppPool
+$IISSiteName = "BookStore"
+$SiteFolderPath = "C:\BookStore"
+$AppPoolName = "BookStoreAppPool"          
+$HostHeader = "www.BookStore.com" 
+
+if(!(Test-Path("$SiteFolderPath"))) {
+    New-Item $SiteFolderPath -type Directory
+    Set-Content $SiteFolderPath\Default.htm "<h1>Hello IIS</h1>"
+}
+
+if(!(Test-Path ("IIS:\AppPools\$AppPoolName"))) {
+    New-WebAppPool -Name $AppPoolName
+}
+
+If((!(Test-Path "IIS:\Sites\$IISSiteName"))){
+    New-Website -Name $IISSiteName -physicalPath $SiteFolderPath  -ApplicationPool $AppPoolName        
+}
+
+if(!(Get-WebBinding -Name $IISSiteName)) {
+    New-WebBinding -Name $IISSiteName -Protocol http -Port 80 -IPAddress "*" -HostHeader $HostHeader
+}
+
+Start-WebSite -Name $IISSiteName
